Move loading controls and rename AsyncActivity to BaseActivity
