+++ app/src/main/res/layout/activity_auth.xml
@@ -60,6 +60,4 @@
             android:layout_marginTop="8dp"
             app:layout_constraintTop_toBottomOf="@+id/image_login_logo"/>
 
-<!--    <include layout="@layout/view_progress"/>-->
-
 </androidx.constraintlayout.widget.ConstraintLayout>
